Tighten up the rustup post
Heavy editing. A lot of it in the intro, but also tightening up the remaining sections and just smoothing the whole thing out a bit.

I lost the rocketships -- they were just a bit too much, with everything else going on in the post.

This is a great post, and I think it flows well.